chore: Fix comment component (#97)
* chore: Add CommentForm component

* chore: Fix css

* chore: Add comment component
